// apps/frontend/lib/generated/schemas/donationsResponseDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Covenant Connect API
 * OpenAPI documentation for the Covenant Connect backend services.
 * OpenAPI spec version: 1.0.0
 */
import type { DonationDto } from './donationDto';

export interface DonationsResponseDto {
  data: DonationDto[];
  page: number;
  pageSize: number;
  total: number;
}