// apps/frontend/lib/generated/schemas/eventsResponseDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Covenant Connect API
 * OpenAPI documentation for the Covenant Connect backend services.
 * OpenAPI spec version: 1.0.0
 */
import type { EventDto } from './eventDto';

export interface EventsResponseDto {
  data: EventDto[];
  page: number;
  pageSize: number;
  total: number;
}